feat: 将 selection 功能代码迁移至业务层
